+++ cmd/gpu_onboard.go
@@ -137,7 +137,10 @@ func runScript(executer Executer, scriptPath string, w io.Writer) error {
 // promptContainer takes container runtime script path as input and prompts the user for confirmation.
 // If confirmed, it runs the script.
 func promptContainer(executer Executer, containerPath string, w io.Writer) error {
-	proceed := utils.PromptYesNo("Do you want to proceed with Container Runtime installation? (y/N)")
+	proceed, err := utils.PromptYesNo("Do you want to proceed with Container Runtime installation? (y/N)")
+	if err != nil {
+		return err
+	}
 	if proceed {
 		err := runScript(executer, containerPath, w)
 		if err != nil {
@@ -153,7 +156,11 @@ func promptContainer(executer Executer, containerPath string, w io.Writer) error
 func promptDriverInstallation(executer Executer, w io.Writer, vendor library.GPUVendor, scriptPath string) error {
 	prompt := fmt.Sprintf("Do you want to proceed with %s driver installation? (y/N)", vendor.String())
 
-	proceed := utils.PromptYesNo(prompt)
+	proceed, err := utils.PromptYesNo(prompt)
+	if err != nil {
+		return err
+	}
+
 	if proceed {
 		err := runScript(executer, scriptPath, w)
 		if err != nil {